benchmark/requirements.txt
fastavro
avro-python3
gitpython
